Move all `TestEnv` to the corresponding `given` packages.
