+++ server/src/test/java/io/spine/examples/chatspn/server/account/given/AccountCreationTestEnv.java
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.examples.chatspn.server.given;
+package io.spine.examples.chatspn.server.account.given;
 
 import io.spine.examples.chatspn.AccountCreationId;
 import io.spine.examples.chatspn.account.ReservedEmail;
