+++ server/src/test/java/io/spine/examples/chatspn/server/message/MessageEditingTest.java
@@ -43,16 +43,16 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
-import static io.spine.examples.chatspn.server.given.MessageEditingTestEnv.editMessageCommand;
-import static io.spine.examples.chatspn.server.given.MessageEditingTestEnv.editMessageCommandWith;
-import static io.spine.examples.chatspn.server.given.MessageEditingTestEnv.messageCannotBeEditedFrom;
-import static io.spine.examples.chatspn.server.given.MessageEditingTestEnv.messageContentCannotBeUpdatedFrom;
-import static io.spine.examples.chatspn.server.given.MessageEditingTestEnv.messageContentUpdatedFrom;
-import static io.spine.examples.chatspn.server.given.MessageEditingTestEnv.messageEditedFrom;
-import static io.spine.examples.chatspn.server.given.MessageEditingTestEnv.messageEditingFailedFrom;
-import static io.spine.examples.chatspn.server.given.MessageEditingTestEnv.messageFrom;
-import static io.spine.examples.chatspn.server.given.MessageTestEnv.createRandomChatIn;
-import static io.spine.examples.chatspn.server.given.MessageTestEnv.sendRandomMessageTo;
+import static io.spine.examples.chatspn.server.message.given.MessageEditingTestEnv.editMessageCommand;
+import static io.spine.examples.chatspn.server.message.given.MessageEditingTestEnv.editMessageCommandWith;
+import static io.spine.examples.chatspn.server.message.given.MessageEditingTestEnv.messageCannotBeEditedFrom;
+import static io.spine.examples.chatspn.server.message.given.MessageEditingTestEnv.messageContentCannotBeUpdatedFrom;
+import static io.spine.examples.chatspn.server.message.given.MessageEditingTestEnv.messageContentUpdatedFrom;
+import static io.spine.examples.chatspn.server.message.given.MessageEditingTestEnv.messageEditedFrom;
+import static io.spine.examples.chatspn.server.message.given.MessageEditingTestEnv.messageEditingFailedFrom;
+import static io.spine.examples.chatspn.server.message.given.MessageEditingTestEnv.messageFrom;
+import static io.spine.examples.chatspn.server.message.given.MessageTestEnv.createRandomChatIn;
+import static io.spine.examples.chatspn.server.message.given.MessageTestEnv.sendRandomMessageTo;
 
 @DisplayName("`MessageEditing` should")
 final class MessageEditingTest extends ContextAwareTest {
