+++ server/src/test/java/io/spine/examples/chatspn/server/message/MessageSendingTest.java
@@ -40,13 +40,13 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
-import static io.spine.examples.chatspn.server.given.MessageSendingTestEnv.messageCannotBeSentFrom;
-import static io.spine.examples.chatspn.server.given.MessageSendingTestEnv.messageFrom;
-import static io.spine.examples.chatspn.server.given.MessageSendingTestEnv.messagePostedFrom;
-import static io.spine.examples.chatspn.server.given.MessageSendingTestEnv.messageSentFrom;
-import static io.spine.examples.chatspn.server.given.MessageSendingTestEnv.randomSendMessageCommand;
-import static io.spine.examples.chatspn.server.given.MessageSendingTestEnv.sendMessageCommandWith;
-import static io.spine.examples.chatspn.server.given.MessageTestEnv.createRandomChatIn;
+import static io.spine.examples.chatspn.server.message.given.MessageSendingTestEnv.messageCannotBeSentFrom;
+import static io.spine.examples.chatspn.server.message.given.MessageSendingTestEnv.messageFrom;
+import static io.spine.examples.chatspn.server.message.given.MessageSendingTestEnv.messagePostedFrom;
+import static io.spine.examples.chatspn.server.message.given.MessageSendingTestEnv.messageSentFrom;
+import static io.spine.examples.chatspn.server.message.given.MessageSendingTestEnv.randomSendMessageCommand;
+import static io.spine.examples.chatspn.server.message.given.MessageSendingTestEnv.sendMessageCommandWith;
+import static io.spine.examples.chatspn.server.message.given.MessageTestEnv.createRandomChatIn;
 
 @DisplayName("`MessageSending` should")
 public final class MessageSendingTest extends ContextAwareTest {
