+++ server/src/test/java/io/spine/examples/chatspn/server/message/given/MessageRemovalTestEnv.java
@@ -24,123 +24,113 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.examples.chatspn.server.given;
+package io.spine.examples.chatspn.server.message.given;
 
 import io.spine.core.UserId;
 import io.spine.examples.chatspn.MessageId;
 import io.spine.examples.chatspn.message.Message;
-import io.spine.examples.chatspn.message.command.EditMessage;
-import io.spine.examples.chatspn.message.event.MessageContentUpdated;
-import io.spine.examples.chatspn.message.event.MessageEdited;
-import io.spine.examples.chatspn.message.event.MessageEditingFailed;
-import io.spine.examples.chatspn.message.rejection.EditingRejections.MessageCannotBeEdited;
-import io.spine.examples.chatspn.message.rejection.EditingRejections.MessageContentCannotBeUpdated;
+import io.spine.examples.chatspn.message.command.RemoveMessage;
+import io.spine.examples.chatspn.message.event.MessageMarkedAsDeleted;
+import io.spine.examples.chatspn.message.event.MessageRemovalFailed;
+import io.spine.examples.chatspn.message.event.MessageRemoved;
+import io.spine.examples.chatspn.message.rejection.RemovalRejections.MessageCannotBeMarkedAsDeleted;
+import io.spine.examples.chatspn.message.rejection.RemovalRejections.MessageCannotBeRemoved;
 
-import static io.spine.testing.TestValues.randomString;
-
-public final class MessageEditingTestEnv {
+public final class MessageRemovalTestEnv {
 
     /**
      * Prevents class instantiation.
      */
-    private MessageEditingTestEnv() {
+    private MessageRemovalTestEnv() {
     }
 
-    public static EditMessage editMessageCommand(Message message) {
-        EditMessage command = EditMessage
+    public static RemoveMessage removeMessageCommand(Message message) {
+        RemoveMessage command = RemoveMessage
                 .newBuilder()
                 .setId(message.getId())
                 .setChat(message.getChat())
                 .setUser(message.getUser())
-                .setSuggestedContent(randomString())
                 .vBuild();
         return command;
     }
 
-    public static EditMessage editMessageCommandWith(Message message, UserId userId) {
-        EditMessage command = EditMessage
+    public static RemoveMessage removeMessageCommandWith(Message message, UserId userId) {
+        RemoveMessage command = RemoveMessage
                 .newBuilder()
                 .setId(message.getId())
                 .setChat(message.getChat())
                 .setUser(userId)
-                .setSuggestedContent(randomString())
                 .vBuild();
         return command;
     }
 
-    public static EditMessage editMessageCommandWith(Message message, MessageId messageId) {
-        EditMessage command = EditMessage
+    public static RemoveMessage removeMessageCommandWith(Message message, MessageId messageId) {
+        RemoveMessage command = RemoveMessage
                 .newBuilder()
                 .setId(messageId)
                 .setChat(message.getChat())
                 .setUser(message.getUser())
-                .setSuggestedContent(randomString())
                 .vBuild();
         return command;
     }
 
-    public static MessageEdited messageEditedFrom(EditMessage c) {
-        MessageEdited event = MessageEdited
+    public static MessageRemoved messageRemovedFrom(RemoveMessage c) {
+        MessageRemoved event = MessageRemoved
                 .newBuilder()
                 .setId(c.getId())
                 .setChat(c.getChat())
                 .setUser(c.getUser())
-                .setContent(c.getSuggestedContent())
                 .vBuild();
         return event;
     }
 
-    public static MessageEditingFailed messageEditingFailedFrom(EditMessage c) {
-        MessageEditingFailed event = MessageEditingFailed
+    public static MessageRemovalFailed messageRemovalFailedFrom(RemoveMessage c) {
+        MessageRemovalFailed event = MessageRemovalFailed
                 .newBuilder()
                 .setId(c.getId())
                 .setChat(c.getChat())
                 .setUser(c.getUser())
-                .setSuggestedContent(c.getSuggestedContent())
                 .vBuild();
         return event;
     }
 
-    public static MessageContentUpdated messageContentUpdatedFrom(EditMessage c) {
-        MessageContentUpdated event = MessageContentUpdated
+    public static MessageMarkedAsDeleted messageMarkedAsDeletedFrom(RemoveMessage c) {
+        MessageMarkedAsDeleted event = MessageMarkedAsDeleted
                 .newBuilder()
                 .setId(c.getId())
                 .setChat(c.getChat())
                 .setUser(c.getUser())
-                .setContent(c.getSuggestedContent())
                 .vBuild();
         return event;
     }
 
-    public static Message messageFrom(EditMessage c) {
+    public static Message messageFrom(RemoveMessage c) {
         Message state = Message
                 .newBuilder()
                 .setId(c.getId())
                 .setChat(c.getChat())
                 .setUser(c.getUser())
-                .setContent(c.getSuggestedContent())
                 .buildPartial();
         return state;
     }
 
-    public static MessageCannotBeEdited messageCannotBeEditedFrom(EditMessage c) {
-        MessageCannotBeEdited rejection = MessageCannotBeEdited
+    public static MessageCannotBeRemoved messageCannotBeRemovedFrom(RemoveMessage c) {
+        MessageCannotBeRemoved rejection = MessageCannotBeRemoved
                 .newBuilder()
                 .setId(c.getId())
                 .setChat(c.getChat())
                 .setUser(c.getUser())
-                .setSuggestedContent(c.getSuggestedContent())
                 .vBuild();
         return rejection;
     }
 
-    public static MessageContentCannotBeUpdated messageContentCannotBeUpdatedFrom(EditMessage c) {
-        MessageContentCannotBeUpdated rejection = MessageContentCannotBeUpdated
+    public static MessageCannotBeMarkedAsDeleted messageCannotBeMarkedAsRemovedFrom(
+            RemoveMessage c) {
+        MessageCannotBeMarkedAsDeleted rejection = MessageCannotBeMarkedAsDeleted
                 .newBuilder()
                 .setId(c.getId())
                 .setChat(c.getChat())
                 .setUser(c.getUser())
-                .setSuggestedContent(c.getSuggestedContent())
                 .vBuild();
         return rejection;
     }
